requirements.txt
streamlit
langchain
langchain_google_genai
langchain_core
python-dotenv
langchain_community
langserve
fastapi
uvicorn